% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.locfit-package}
\alias{mlr3learners.locfit}
\alias{mlr3learners.locfit-package}
\title{mlr3learners.locfit: Learners from the {locfit} package for 'mlr3'}
\description{
Adds learners from the {locfit} package to 'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
